requirements.txt
telethon
load_dotenv
pysocks
